Functional
To turn ins as is, this program requires additional commenting. I don't
fancy that I have to create an object of my CountCharacter class to use
it. I would like to explore how to implement as its own namespace or
some such, as I have read online this make allow me to forgo the
aforementioned "issue"; and, I have also read that creating static
classes/methods in C++ is not generally a good practice