Working except for exception handling
Converted CountCharacter class to be static after speaking with
instructor; and, the instructor also stated that we should include a
try...catch block to get familiarity with exception handling. I have
attempted to define my own exception handler, adapted from pg 607+ of
text; but it thus far does not work.